backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.28
chromadb==0.5.11
duckduckgo-search==6.3.0
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
sqlalchemy==2.0.35
aiosqlite==0.20.0